mixin for UUID generation
